TS version of #470 (Refactor GitHub client extensions)
